Registration System
This directory consists of forms that were used to fill in the details
of participants attending the event. Also one could check details of
their record by using the unique ID(barcode_id) which was given to them
in form of a Registration Card. This System was mainly used by the
logistics team.